Projects/Movie_reviews_sentiment_analysis/requirements.txt
bs4
click==8.0.3
fastapi==0.70.0
flake8==4.0.1
matplotlib==3.4.3
nltk==3.6.5
numpy==1.21.2
pandas==1.3.3
pyyaml==6.0
requests==2.25.1
scikit-learn==0.24.2
seaborn==0.11.2
uvicorn==0.15.0